Minor module refactor in core/renderer/ (#10643)
* Move render-scene and native-scene

* Minor changes
