+++ cocos/3d/models/skinning-model.ts
@@ -23,8 +23,6 @@
  THE SOFTWARE.
 */
 
-
-
 import { JSB } from 'internal:constants';
 import { Material } from '../../core/assets/material';
 import { RenderingSubMesh } from '../../core/assets/rendering-sub-mesh';
@@ -39,8 +37,7 @@ import { ModelType } from '../../core/renderer/scene/model';
 import { uploadJointData } from '../skeletal-animation/skeletal-animation-utils';
 import { MorphModel } from './morph-model';
 import { deleteTransform, getTransform, getWorldMatrix, IJointTransform } from '../../core/animation/skeletal-animation-utils';
-import { BatchingSchemes, IMacroPatch, Pass } from '../../core/renderer';
-import { NativeJointInfo, NativeJointTransform, NativeSkinningModel } from '../../core/renderer/scene';
+import { BatchingSchemes, IMacroPatch, Pass, NativeJointInfo, NativeJointTransform, NativeSkinningModel  } from '../../core/renderer';
 import { warnID } from '../../core/platform/debug';
 
 const myPatches: IMacroPatch[] = [
