+++ cocos/core/renderer/core/render-scene.ts
@@ -25,14 +25,14 @@
 import { JSB } from 'internal:constants';
 import { Root } from '../../root';
 import { Node } from '../../scene-graph';
-import { Camera } from './camera';
-import { DirectionalLight } from './directional-light';
-import { Model, ModelType } from './model';
-import { SphereLight } from './sphere-light';
-import { SpotLight } from './spot-light';
+import { Camera } from '../scene/camera';
+import { DirectionalLight } from '../scene/directional-light';
+import { Model, ModelType } from '../scene/model';
+import { SphereLight } from '../scene/sphere-light';
+import { SpotLight } from '../scene/spot-light';
 import { TransformBit } from '../../scene-graph/node-enum';
 import { DrawBatch2D } from '../../../2d/renderer/draw-batch';
-import { NativeDrawBatch2D, NativeRenderScene } from './native-scene';
+import { NativeDrawBatch2D, NativeRenderScene } from '../native-scene';
 
 export interface IRenderSceneInfo {
     name: string;
