+++ cocos/core/renderer/index.ts
@@ -27,8 +27,6 @@ import config from './config';
 
 import * as scene from './scene';
 
-import './scene/deprecated';
-
 export { createIA } from './utils';
 
 const addStage = config.addStage;
@@ -42,4 +40,8 @@ export * from './core/texture-buffer-pool';
 export * from './core/material-instance';
 export * from './core/pass-instance';
 export * from './core/memory-pools';
+export * from './core/render-scene';
+export * from './native-scene';
+export * from './deprecated';
+
 export { scene };
