+++ cocos/core/renderer/scene/camera.ts
@@ -24,18 +24,16 @@
  */
 
 import { EDITOR, JSB } from 'internal:constants';
-import { enums, Frustum, Ray } from '../../geometry';
+import { Frustum, Ray } from '../../geometry';
 import { SurfaceTransform, ClearFlagBit, Device, Color, ClearFlags } from '../../gfx';
-import {
-    lerp, Mat4, Rect, toRadian, Vec3, IVec4Like,
-} from '../../math';
+import { lerp, Mat4, Rect, toRadian, Vec3, IVec4Like } from '../../math';
 import { CAMERA_DEFAULT_MASK } from '../../pipeline/define';
 import { Node } from '../../scene-graph';
-import { RenderScene } from './render-scene';
+import { RenderScene } from '../core/render-scene';
 import { legacyCC } from '../../global-exports';
 import { RenderWindow } from '../core/render-window';
 import { preTransforms } from '../../math/mat4';
-import { NativeCamera } from './native-scene';
+import { NativeCamera } from '../native-scene';
 import { warnID } from '../../platform/debug';
 
 export enum CameraFOVAxis {
