+++ cocos/core/renderer/scene/light.ts
@@ -26,9 +26,9 @@
 import { JSB } from 'internal:constants';
 import { Vec3 } from '../../math';
 import { TransformBit } from '../../scene-graph/node-enum';
-import { RenderScene } from './render-scene';
+import { RenderScene } from '../core/render-scene';
 import { Node } from '../../scene-graph';
-import { NativeDirectionalLight, NativeLight, NativeSphereLight, NativeSpotLight } from './native-scene';
+import { NativeDirectionalLight, NativeLight, NativeSphereLight, NativeSpotLight } from '../native-scene';
 
 // Color temperature (in Kelvin) to RGB
 export function ColorTemperatureToRGB (rgb: Vec3, kelvin: number) {
