+++ cocos/core/renderer/scene/shadows.ts
@@ -29,9 +29,9 @@ import { Sphere } from '../../geometry';
 import { Color, Mat4, Vec3, Vec2 } from '../../math';
 import { legacyCC } from '../../global-exports';
 import { Enum } from '../../value-types';
-import { ShadowsInfo } from '../../scene-graph/scene-globals';
+import type { ShadowsInfo } from '../../scene-graph/scene-globals';
 import { IMacroPatch } from '../core/pass';
-import { NativeShadow } from './native-scene';
+import { NativeShadow } from '../native-scene';
 import { Shader } from '../../gfx';
 
 /**
