+++ cocos/core/renderer/scene/skybox.ts
@@ -32,9 +32,9 @@ import { UNIFORM_ENVIRONMENT_BINDING, UNIFORM_DIFFUSEMAP_BINDING } from '../../p
 import { MaterialInstance } from '../core/material-instance';
 import { Model } from './model';
 import { legacyCC } from '../../global-exports';
-import { SkyboxInfo } from '../../scene-graph/scene-globals';
+import type { SkyboxInfo } from '../../scene-graph/scene-globals';
 import { Root } from '../../root';
-import { NaitveSkybox } from './native-scene';
+import { NaitveSkybox } from '../native-scene';
 import { GlobalDSManager } from '../../pipeline/global-descriptor-set-manager';
 import { Device } from '../../gfx';
 import { Enum } from '../../value-types';
