+++ cocos/core/renderer/scene/submodel.ts
@@ -33,7 +33,7 @@ import { legacyCC } from '../../global-exports';
 import { ForwardPipeline } from '../../pipeline';
 import { errorID } from '../../platform/debug';
 import { Shadows } from './shadows';
-import { NativePass, NativeSubModel } from './native-scene';
+import { NativePass, NativeSubModel } from '../native-scene';
 import { getPhaseID } from '../../pipeline/pass-phase';
 import { Root } from '../../root';
