+++ cocos/core/renderer/utils.ts
@@ -23,8 +23,6 @@
  THE SOFTWARE.
  */
 
-
-
 import { Attribute, Buffer, BufferInfo, Device, InputAssemblerInfo, AttributeName, BufferUsageBit, Format, MemoryUsageBit } from '../gfx';
 import { IGeometry } from '../../primitive/define';
