+++ cocos/core/root.ts
@@ -32,10 +32,11 @@ import { JSB } from 'internal:constants';
 import { builtinResMgr } from './builtin';
 import { Pool } from './memop';
 import { RenderPipeline, createDefaultPipeline, DeferredPipeline } from './pipeline';
-import { Camera, Light, Model, NativeRoot } from './renderer/scene';
+import { Camera, Light, Model } from './renderer/scene';
+import { NativeRoot } from './renderer/native-scene';
 import type { DataPoolManager } from '../3d/skeletal-animation/data-pool-manager';
 import { LightType } from './renderer/scene/light';
-import { IRenderSceneInfo, RenderScene } from './renderer/scene/render-scene';
+import { IRenderSceneInfo, RenderScene } from './renderer/core/render-scene';
 import { SphereLight } from './renderer/scene/sphere-light';
 import { SpotLight } from './renderer/scene/spot-light';
 import { legacyCC } from './global-exports';
